Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)
> 
> Regarding a possible default value for the regex (not saying this is dearly need ATM): By "the namespace *name*" you mean e.g. "Item" (then followed by the colon)? Is that not in `repoNamespaces`?

Hm, I guess it is. (It’s a callback and never evaluated during normal operation, so I didn’t see it when searching for Item or Property in a var_dump() of a WikibaseClient instance.) But it doesn’t seem to be used anywhere at all (Wikibase Code Search [1] only finds definitions, not accesses), and the default value in WikibaseClient.default.php warns that it doesn’t match the example settings (assumes items are in the main namespace)). It looks like one of the last uses was removed two years ago in T153769, which also mentions that we should “check if we can drop some of the client settings, such as 'repoNamespaces' or 'repoArticlePath'.” So I’m not sure if we really want to rely on this value.

[1]: https://codesearch.wmflabs.org/search/?q=repoNamespaces

Patch-set: 3
Reviewer: Lucas Werkmeister (WMDE) <4699@e9e9afe9-4712-486d-8885-f54b72dd1951>