Update patch set 2
Patch Set 2: Code-Review+1

I feel like this should be more configurable per entity type… the reason to disallow moves isn’t directly whether there is an entity in the main slot, the reason (I think) is that for Items, Properties and Lexemes the entity ID is tied to the page title, so moving the page would invalidate the entity ID. For MediaInfo entities, the entity ID is tied only to the page ID, so moves are fine. But there could also be hypothetical other entity types where the entity ID is tied to the page title even though the entity doesn’t occupy the main slot, or vice versa.

That said – this kind of hypothetical concern definitely shouldn’t block this fix for a pretty serious issue on Commons, and I think it’s totally fine to merge this for now.

Patch-set: 2
Reviewer: Lucas Werkmeister (WMDE) <4699@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1
Work-in-progress: false