bridge: conditionally load module
Currently, anyone can load our work-in-progress data bridge code in
production by manually loading the wikibase.client.data-bridge.init
ResourceLoader module – the dataBridgeEnabled setting only controls
whether we load the module by default or not, it’s always available.
This is not optimal. (Also, having an unneeded module in the dependency
graph slightly hurts performance.)

ResourceLoader won’t let us completely remove a module later, and we
can’t conditionally add it to the array in Resources.php because at this
point we haven’t even loaded the default client settings yet, let alone
LocalSettings.php, but what we can do is create the module via a factory
callback (which is called late enough that we have access to settings)
and conditionally clear its list of targets. This ensures that, if data
bridge functionality is not enabled in the configuration, the module is
not available at all. (This trick was suggested by Pablo Grass.)

Bug: T226982
Change-Id: If147649bde9c7d7676d74960e1ed91be01c1b271