bridge: don't forbid es6 methods in eslint
Upstream Wikimedia eslint config uses `no-restricted-properties` and
`no-restricted-syntax` to disallow some methods not allowed in the
version of the standard we are targeting. However, as part of our build
process we either polyfill or transpile away those methods, so we’re not
interested in that part of the config; we only want the “rules” part,
which since v0.13.0 we can import via a separate file.

Unfortunately, this requires us to drop the client import, since it
pulls in es5 (and not just the “rules” part of it), again including the
`no-restricted-syntax` checks that we don’t want to use. For now, we can
copy three client-specific rules and one env to our own config – perhaps
in future the upstream client config can be split up suitably.

Bug: T227318
Change-Id: I9fd963a1d2f70c387ce380c3231f958cd3087181