bridge: export app function and use in init
At some point the init module will call into the app module to
initialize itself with the information from the edit links; for now, we
still just log something to the console, but this establishes the module
communication process (previously, the module would only be loaded and
run its own top-level code).

To make the app module’s exports visible to the init module via
ResourceLoader, we have to build the app in library mode. All credit for
this goes to Pablo – I had a much hackier solution for this.

This facilitates two separate calls to build.
Attempts to have two different entry points, one of them packaging app
as lib, using [0] to work around [1], were not successful.

[0]: https://github.com/jkzing/vue-cli-plugin-multi-compiler
[1]: https://github.com/vuejs/vue-cli/issues/2993

See also
https://cli.vuejs.org/guide/build-targets.html#library
https://webpack.js.org/guides/author-libraries/
https://webpack.js.org/configuration/configuration-types/#exporting-multiple-configurations

This also adds a test that the init module actually calls the app
module; to make this possible, the init module has to return a Promise
that resolves after it’s done, so that the test can wait until then to
run its assertions. This is implemented using async/await syntax – of
course, we could also use Promise.then() “by hand”.

Bug: T227178
Change-Id: Ic2714b58493641bc5939a5c802f83594b33e1a0c