Add FallbackLabelDescriptionLookup interface
Due to the way it’s implemented, since change I1ba2142d72 (commit
b0d02fbaf5) WikibaseLanguageDependentLuaBindings has required a
label/description lookup that returns specifically instances of
TermFallback, not just Term as guaranteed by the LabelDescriptionLookup
interface. It used to require LanguageFallbackLabelDescriptionLookup for
this, but in change I75b4c519c4 (commit bb9aca3238) this was changed
back to the LabelDescriptionLookup interface so that an instance of
UsageTrackingLanguageFallbackLabelDescriptionLookup could be injected –
however, this makes Phan unhappy because we call the TermFallback method
getActualLanguageCode() on objects that might be any kind of Term. To
resolve this, introduce a new interface FallbackLabelDescriptionLookup
that’s implemented by both LanguageFallbackLabelDescriptionLookup and
UsageTrackingLanguageFallbackLabelDescriptionLookup, and require that
interface in WikibaseLanguageDependentLuaBindings.

Change-Id: Ie39f06bb334ed191ac4aea1126b54bf8c61119c1