bridge: use OO dialog and mount vue app into it
This creates a OO dialog[0] with some basic content and an identifier
which we then use to mount our Vue application inside of.
Using the very OO component ensures future alignment with possible
developments of this component and focuses bridge development on the
inside of this dialog. As OO itself is not written in typescript and
there appear to be no type definitions this patch adds a lot of
interface definitions (MwWindow) and, in the tests, even pseudo-code
where we simulate the presence of the OO base classes to arrive at
running code.

Passes the container selector (#data-bridge-container) to the app
for mounting. See [1]

I found [2] helpful to wrap my head around how TS & JS play together
in this patch but it still leaves things to be desired.

[0] https://doc.wikimedia.org/oojs-ui/master/js/#!/api/OO.ui.Dialog
[1] https://vuejs.org/v2/api/#vm-mount
[2] https://www.sitepen.com/blog/advanced-typescript-concepts-classes-and-types/#highlighter_43462

Bug: T227179
Change-Id: I2129e8fe584f80fbfcef7cb354bbac4deb607d5d