lib: make vue 2 available as ResourceLoader module
Make vue 2 available as module through ResourceLoader to allow
use of it from other modules so that that they don't ship it
along with their own code.
This reduces redundant payloads and allows for "application
libraries" in which vue gets "externalized" during the build process.
See
https://cli.vuejs.org/guide/build-targets.html#library

Naming of this package was based on the semver logic that modules
depending on the MAJOR 2 release of vue should be fine™ with any MINOR
or PATCH release of it, but we'll need another module for vue 3, ...

This also makes bridge's dependency on this, which is packed as lib
since Ic2714b58493641bc5939a5c802f83594b33e1a0c, explicit. (So far there
is no real application code so no loss of feature could be witnessed.)

You can try this by, in your browser console, running
	mediaWiki.loader.using( 'vue2' ).then( function( require ){
		var Vue = require( 'vue2' ); // this is what bridge packed as lib will do
		console.log( Vue.config );
	} );

Bug: T227179
Change-Id: Idbc9cf2fe039564873d75ab4c42ca7be9f037ca1