bridge: Add component to display edit data
The component itself is obviously just a placeholder, to be moved and
changed in the next commit.

This sets the dimensions of our modal on the top-level component. They
are hard-coded for now. At some point we want to do a commit that sets
up our styling system consistently and with proper variable usage.

The class prefix scheme is intended to be used for data-bridge specific
components. It is chosen in a way that follows the basic idea of
termbox, while still being specific for data-bridge. The more general
`wb-ui-` prefix remains usable for more generic, reusable components.

Bug: T227180
Change-Id: I14904c08c45eac67c835836f28a5eb6f9be798f2