Fix namespace of MockClientStore
This class is also used in ArticlePlaceholder, so we need a temporary
compatibity class_alias(). However, we should be able to remove it soon.

For some reason, MediaWiki core’s AutoLoaderStructureTest expects the
new class name to be explicitly listed in the $wgAutoloadClasses even
though it can be autoloaded via PSR-4, so we temporarily have an
autoload.php that is not *quite* generated by generateAutoload.php. This
will be resolved once we remove the compatibility alias, though.

Bug: T172368
Change-Id: I372453aad0eacf3060674dc635cc23da0e8bbe96