bridge: add getter for main value to edit
This connects the entity store’s getter for a value (added in change
Ib0e9230187) with the root store’s getter for the property ID to be
edited by this data bridge instance. Since it inherits the restrictions
of the entity store’s getter (only supports the simplest case, throws
errors otherwise), it should likewise be considered rudimentary for now.

Bug: T227187
Change-Id: Ia388351a2f01501738e607edb1bf0bad9b657209