bridge: add browser tests
Introduces a basic browser test to make sure that integration with
real OO actually works.
They are not run as part of CI yet as core tests/selenium/wdio.conf.js
does not discover out spec (yet). For the time being you can add the
following line there to execute it
        relPath( './extensions/Wikibase/client/data-bridge/tests/selenium/specs/**/*.js' ),
The spec assumes that dataBridgeHrefRegExp points to beta wikidata.

Bug: T227179
Change-Id: Ib07eb233e1c38b7cbbb7520a65ce6980f8fd7cec