bridge: add WbRepo interface for wbRepo config value
The init module will use this value soon to discover the location of
Special:EntityData – this commit adds an interface for it, overloads
mw.config.get() to return it, and sets up the tests to return this value
as well (previously they could only return wbDataBridgeConfig).

To avoid repetition in the tests, the mwMockEnv() function (which
already existed in two very similar versions) is extracted into a new
file, accompanied by a new mockMwConfig() function. (This also requires
that MwWindow exports the MwConfig interface.)

Bug: T228408
Change-Id: I94921ace2ad261f39f8a406910f030dbab496fce