bridge: add dev entry point
This patch introduces the dev-box (npm run serve).
In terms of that, launch and main.ts is merged in mock-entry. We should
discuss if we want to fill servcies/hand over via launch. This would
establish a very simple rule to mark something as services - everything
which is shared by app and init is a service, therefore app must be
agnostic towards mediawiki.

Bug: T227192
Change-Id: I58e22904cf6d3fd7cf74a6425eec7a7d559cbe51