bridge: bind store as the only source of truth
This patch wires up the store with the components. Therefore it makes
significant changes in Terms of "the store is the only source of truth"
principle. It wires up with the ApplicationInformationRepository as well
to pass base application information into the store instead of using
properties at the root component. (There was also a different approach
see: I4e0bd1e16cea616f62591fda21266a8ee5d04846)

Bug: T229109
Change-Id: I9206c785fe78527af76b58ed7746f458972176e2