Wrap Vue2 JS file in IIFE
vue2.common.prod.js pollutes the global environment and overwrites the
standard `$` (jQuery) variable. ResourceLoader usually protects against
this by wrapping each module in an immediately invoked function
expression, but that doesn’t happen in debug mode; to avoid Vue breaking
the page, we can wrap it in an IIFE of our own.

Bug: T229390
Change-Id: Idf6c7ee0dcef9eae06ba6547658e497acea5fd4c