bridge: use wdio-wikibase to create local items
This makes the tests more portable and they require less setup for the
local wiki in order to run.

I have chosen to not abstract the createItem call yet, as I find it to
be not yet clear what future requirements are going to be. A good moment
to do that would be when we added the first test outside of init.js.

Bug: T229444
Change-Id: I74e69765f039c1d3d0e0cc07de539cd09988a486