bridge: migrate to webdriverio 5
Version 5 is the current version of webdriverio. There is already a task
to switch mediawiki core tests to this new version: T213268.  At some
point we will likely want to switch Wikibase browser tests to version 5
as well. So the intention of this patch is to not write even more tests
with the old version's syntax.

Once Wikibase has been migrated to webdriverio version 5, we might want
to consider integrating back into it.

However, webdriverio 5 is a breaking change in some relevant aspects.
Two aspects are especially important to point out here:

Protocol commands now return the value directly, instead of an object
with the value being on its value property. This has the consequences
that we have to keep ForwardCompatUtil, until wdio-mediawiki updates.

Custom configurations options set in the wdio.conf.js are now available
in browser.config instead of browser.options. This has the consequence
that we have to add username and password to calls to the mediawiki API
and that storing the screenshots currently doesn't work.

As it will take some time for these changes to be implemented in
wdio-mediawiki and then to be released on npm, the affected methods have
been copied here for now and adjusted.

Change-Id: I064f094c199fae7be52ac69854b386d315497e36