Run data-bridge browser tests in CI
Introduces a WikibaseClient.ci.php for settings to be used in testing
which we don't yet consider ready for the default example at the
moment.

This uses npm-run-all in order to run the browser tests for
wikidata-bridge at the same time, but independently from the browser
tests of Wikibase repo.

The option --continue-on-error ensures that the tests for data-bridge
are run even if there are failures in the tests of Wikibase repo. If
there are failures in any of the commands, the exit-code of the parent
npm-run-all command will always be non-zero.
From the documentation of that option:

> Set the flag to continue executing
> other/subsequent tasks even if a task threw an
> error. 'npm-run-all' itself will exit with
> non-zero code if one or more tasks threw error(s)

Bug: T228523
Change-Id: Iae49b96fef4bf0270ee26ac88fc686d1a876785a