Add $tags parameter to EntityStore, EditEntity
This makes it possible to set change tags on the MediaWiki revision
corresponding to an edit. Callers are responsible for checking that the
user is allowed to set these tags, since they are better suited for
dealing with the permission error that otherwise results.

This includes an incompatible change to the EntityStore interface (extra
parameter to saveEntity()), so to avoid “incompatible definition”
errors, EntityStore implementations in WikibaseLexeme are updated first.

EntityStore::saveRedirect() should, in principle, also have a $tags
parameter, but that’s left for later. (A corresponding TODO is added in
WikiPageEntityStore::saveRedirect(), at least.)

Bug: T229917
Change-Id: I24782e47b95ab2c909abc7d0a350a08e7f4a0dda
Depends-On: I5a717705a5ecc18bff6d5797e2eba5aff10fdaa1
Depends-On: Id8f14044ba5d926447aedd60e8cc8eb7dc864899