Add tags parameter to editing API modules
Most of the API modules that edit an entity gain a 'tags' parameter to
set the tags of the new revision. It’s handled in the same place as the
common 'bot', 'baserevid' and 'token' parameters, so there is no
duplicate code (though there are duplicate i18n messages).

Note that this does not affect modules that create redirects
(wbcreateredirect, wbmergeitems) or modules in WikibaseLexeme, neither
of which use EntitySavingHelper or ModifyEntity.

Bug: T229917
Change-Id: Ic96e1b1e674d7a2fd774521456e6a92bf333af5a