bridge: integrate StringDataValue component into Bridge component
This injects a setter callback into the StringDataValue component,
because that component shouldn't know where that data came from or how
it gets back there again.

The component that orchestrates this is currently the Bridge component.
While that might not be ideal, this will move as the app as a whole
evolves.

This commit makes the DataPlaceholder component obsolete.

Further, it removes the `text-align: center` from the App component as
this was causing trouble with the label. Apparently, that rule was still
left-over from the initial commit of the vue-js auto-generated
boilerplate.

I have no idea why typescript started to complain about the line
```
return this.dataValue.value;
```
as I didn't touch it at all and it apparently was not an issue in the
previous commit. But the added explicit cast to string made things quite
again.

Bug: T228263
Change-Id: I3f6226cf99483985f454c0ee991e262cfae5b98b