bridge: Add Header component
Relevant ooui styles and inspiration have been taken from:
https://github.com/wikimedia/oojs-ui/blob/d3cefbbc4fa708069f6b86008f02e08e37e36c88/src/themes/wikimediaui/windows.less#L115
https://www.mediawiki.org/wiki/OOUI/Windows/Process_Dialogs
https://doc.wikimedia.org/oojs-ui/master/demos/?page=dialogs&theme=wikimediaui&direction=ltr&platform=desktop

This uses a <div> instead of a semantic <header> tag because the claim
has been made that a <header> should not contain navigational elements.

The cyan background-color in the stories has been added in order to show
that the buttons switched sides in the rtl-story for people who do not
read Hebrew.

Bug: T230328
Change-Id: I6fb4a5a9bc6d0767d72ccb7698d9898a62c32eed