bridge: fix flaky bridge browser test
This test failed both with the error that the waitUntil for the input
focus timed out and the error that apparently no keys were added to the
string.

Moving the clicking of the input into the waitUntil, hopefully resolves
the problem of it timing out.

Wrapping the `browser.keys` call into `browser.call` hopefully resolves
the issue that the following assert happens before `browser.keys` is
done typing.

Bug: T231074
Change-Id: Ia3410407ebbc284f5f285f94a525baaa9459570b