bridge: use padding-less panel for bridge dialog
In order to be in full control of how the inside of our bridge dialog
looks like, remove the default padding and apply it to the inside of our
application instead.
See https://doc.wikimedia.org/oojs-ui/master/js/#!/api/OO.ui.PanelLayout

This aligns the $base-font-size-* variables with the ones seen in
common.less but mends their units to better work with SASS. A helper
function allows us to quickly "convert" px measurements to em.

Changes the dialog header min-height and font-size, calculation of which
are based on this variabes, to the correct values in the process.
See common.less lines 58 and 191 respectively.

Also mends the dev view (mock-entry) template to properly simulate the
integration with real OOUI, incl the font size we inherit there.

Bug: T230328
Change-Id: I9ac423a622e5c641ea5d6233ec56bf8cb792212e