bridge: try to fix flaky test again
Frustratingly, the test still fails in verification because the assert for
the added characters is still happening before `browser.keys()` is done.
This commit now tries to explicitly wait for the content of the input
area to be as expected instead of formally asserting it.

This isn't nice, but maybe it works.

See also commit 17b617068277ad157bcd03aedf97c6940c54793a

Bug: T231074
Change-Id: I6eb5cbae7ae21ef6b60b5eec019b1a9ba81ef95d