Correct languages in EntitySourceDocumentUrlProviderTest
The old version of the first test only worked if the wiki’s content
language was English, as it assumed the special page URL would use the
Special: namespace. Instead, set the content language to German in both
tests, and assert the German namespace name in the first test, and the
canonical namespace name in the second one.

Bug: T231666
Change-Id: Ia4d0088ffbfafcb3c7c3ca35995615b0f2771663