bridge: Add interface for EntityLabelRepository
In order to keep the interface specific, this Repository returns only
the information it was asked for, without additional boilerplate. It is
up to the client to decide what to do with that information and what
boilerplate is desired for storing it.

The language to use in this repository is expected to be provided in
the constructor.

This copies the Term interface from Termbox.

Bug: T231465
Change-Id: I6cafce2dc7f630fe2285bff7254745ef3fedabb4