bridge: fix types of newMockStore
The context that is passed into a action is not the same as the one with
which a store is constructed – the action expects the getters to be
simple values, not getter definitions. The old type only worked because
in the existing tests all those values are callable, so they also
satisfied the Getter type.

Change-Id: I56f5a2db97059644bd1be247a2edca0295f07281