Remove most uses of call_user_func_array()
In most cases, we can use the PHP5.6+ spread syntax instead, which in my
opinion is more readable, and also makes it easier to find/navigate all
uses of a method (instead of hiding some of them behind a string
literal).

In SetClaimValueTest::testValidRequests(), some string keys need to be
removed from the argument arrays because call_user_func_array() allows
them but ... does not. However, I think the values are sufficiently
clear that it’s not necessary to document the parameter names in
comments instead.

MagicWordHookHandlersTest still needs call_user_func_array() because
TestingAccessWrapper otherwise loses the referenceness of $ret.

Change-Id: I4fb08ebbdcf3d918e34a24bd99941974490ff886