Remove most uses of func_get_args()
We can use PHP5.6+ variable arguments syntax instead, which in my
opinion is more readable and also in some cases removes the need to
array_shift() away other arguments first.

Some of the functions touched here also support an array in place of the
first argument. In the long run, I would like to get rid of that – I
assume it was only ever a convenience feature, which with “...” syntax
should no longer be necessary – but for now we continue to support it.

Change-Id: I3cba7f06437104ef4acbfd94928ba103c4e37d1d