PSR-4 autoload more of Wikibase Client
The only obstacle to this is the EntityUsage class, whose constants we
already need in WikibaseClient.default.php – however, at that point, the
AutoloadNamespaces from extension-client-wip.json haven’t been processed
yet (the wfLoadExtension() call in WikibaseClient.php only enqueues the
file for later processing). One simple workaround for that is to
manually load the class, and for a single class I think that’s
acceptable. (WikibaseClient.default.php uses a handful of other classes,
but all of them are only referenced in callbacks – by the time those are
called, autoloading is fully set up.)

Bug: T88258
Change-Id: Ie8a77230bacf5febdc0078e4734b471f177a047a