Move WikibaseView-only modules from lib/resource/lib to WikibaseView
This partially reverts parts of I0e190b0c72 but for good reason
most of the moved modules are not needed in lib or client. They
are only used in WikibaseView/WikibaseRepo

We haven't removed the files from wikibase data-values value-view lib.
So removing the files would be enough. They are in the submodule.

Bug: T203696
Change-Id: Id7708953f16e64e60d9741a5afa4e0da7795497e