bridge: External Events
This patch introduces and wires the event driven approach to handle
ooui dialog close.
Events are directly wired and repeated from app to init. Though it repeats only only a certain amout of events, which are defined in src/events/index.ts.
Please note, the repeated events could also be used in
vue, if there is any collision by naming.

Bug: T230337
Change-Id: I3bf6f463ddf49b4a599b23a085df9b4b248aa111