Use qqx language in SpecialEntityUsageTest
This makes the test independent of the wiki’s content language, which
might help with performance a bit.

I believe the first test was also broken on wikis in languages like
Chinese, where the colon-separator message uses a fullwidth colon
instead of a regular colon plus a space, or French, where colons and
some other punctuation are preceded by a non-breaking space.

This is also a dependency of I67632cc226.

Change-Id: I3d6cba177d3321e8703fa29b607fda29c6bd599a