bridge: Make EventEmittingButton accessible if used as button
This makes the EventEmittingButton more accessible when used as a button
and at the same time maintains its functionality to be used as a link.

A link with a href-attribute will still show up in a list of links as
created by assistive software. Also, it still offers to open the "page"
in a new tab in the context menu and it responds to middle-click
accordingly.

A link without href is usually not focusable via keyboard.
`tabindex="0"` is added to remedy this. The 0 says to keep this element
in the normal tab-order.
The `role="button"` will make the element appear as a button to the
screenreader.
The keydown and keyup-handler are added to actually respond to the click
and to be able to style it. This implements the OOUI behavior for
"buttons" without href, while it maintains the EventEmittingButton's link
behavior if there is a href.

If there is a link, this keeps the spacebar behavior unaltered as that
is the current behavior in Termbox and is simplest for now.

Bug: T230326
Change-Id: I18130c156c61fd4e30b3a7385c6f1cae111ac9a2