Stop using deprecated function wfWaitForSlaves()
Instead, inject an ILBFactory into the four helper classes that
previously called the function, and call waitForReplication() on that,
with parameters equivalent to the previous wfWaitForSlaves() call. (In
ChangePruner, this also replaces uses of wfGetDB() with proper
dependency injection; EntityUsageTableBuilder already had an
ILoadBalancer injected, which is now replaced with an ILBFactory.)

Note that wfWaitForSlaves() has different default behavior from
waitForReplication() – unless otherwise specified, it waits for the
current wiki’s main cluster, whereas waitForReplication() by default
waits for all clusters used in this request. I’m not sure if this
behavior is actually intended by the three callers that didn’t supply
the $wiki argument for wfWaitForSlaves(), but I’ve preserved it for now.

Bug: T181673
Change-Id: I749c10b7cced30a6c7e03f3165dbab56c45c7ceb