bridge: separate configurations of test types
Separate configuration for unit and integration tests.

Apparently one can not combine testMatch and a matcher (e.g. to a
specific file) passed as unnamed argument and expect results to match
one AND the other.

This allows for execution of all of the following options
* npm run test:unit
* npm run test:integration
* npm run test:unit tests/unit/mediawiki/createServices.spec.ts
* npm run test:integration tests/integration/store/actions.spec.ts

It also cleans up the package.json a bit so the extra files may just
be worth it.

Funny syntax (modifying the export) in the configs is a way of not
colliding with filenames/match-exported which otherwise complains.

Change-Id: I925622609f0c6132684eb6a5a6425f378a8f15ac