Use LoadBalancerSingle instead of FakeLoadBalancer where possible
LoadBalancerSingle used to close its injected connection, which made it
unsuitable for most tests (where $this->db needs to be kept open for
teardown) and prompted us to write FakeLoadBalancer. This was changed
with MediaWiki core change I61914b9866 (commit 143333b172), though, so
now we can use LoadBalancerSingle instead of FakeLoadBalancer in most
tests.

A few tests are excluded, for two possible reasons:
* they actually inject different read and write connections into
  FakeLoadBalancer, a feature not supported by LoadBalancerSingle; or
* they don’t actually have a database, and FakeLoadBalancer accepts null
  connections in the constructor but LoadBalancerSingle does not.

Change-Id: Ib833f2d1b90a2ea3ae39ccd389ee12064766919c