bridge: prevent flaky browser test when closing app
Occasionally the !DataBridgePage.app.isExisting(); timed out in jenkins
testing. Notably, the screenshot that is created on failure no longer
showed the app.
This patch now waits first for the app to no longer being visually
displayed before waiting for it to be removed from the DOM.
This could tell us multiple things:
* if the isDisplayed check becomes flaky, then either our closing isn't
reliable or the test is somehow broken (looking at the wrong element?)
* if the isExisting check keeps failing, then either the elements are
not reliably removed from the DOM or the way how we select for the
elements in the DOM is broken
* if this failure no longer occurs, then we might just need more time

An alternative/additional measure would be to add video recording to our
tests, similar to the ffmpeg recording in core. This might be added in a
future patch.

Bug: T233830
Change-Id: I033167c56951ce9ee50e03f9616dd2635b652bde