Use LBFactorySingle instead of FakeLBFactory where possible
This is not quite a straightforward replacement, since LBFactorySingle
wraps a single *connection* (an intermediate LoadBalancerSingle is
instantiated by the class itself), not a single *load balancer* (of any
kind). But in cases where the load balancer we used was a
LoadBalancerSingle anyways, we can simplify by using LBFactorySingle
(and, where we need an ILoadBalancer as well, getting it back out of the
load balancer factory).

Change-Id: Ida1fd3671708149733ac1ea1efc11ded5df3a945
Depends-On: I6cb62cfeef299b773cf62c16aa4d3d6f0496b7fe