Document requirement for tag in test README
The browser test for tagging edits requires that the data bridge is
configured to tag its edits with “Data Bridge” (it creates the tag if
necessary, but there’s no good way for it to overwrite the data bridge
config, since init.ts reads it as soon as it’s loaded).

While we’re at it, also fix WikibaseClient.ci.php, which accidentally
sets the edit tags to a single string instead of an array of strings
(this only worked accidentally – the string made it all the way through
wbDataBridgeConfig and ForeignApiWritingRepository to mw.ForeignApi,
which accepts string and array parameters alike).

Change-Id: I03db7d0cadd75f2621a3730a55711cf356c44002