bridge: add cancel button
Adds the cancel/close button, largely inspired by termbox.
The RTL question, in the hope we can find a clean solution on the
application level and do not have to solve this within the
individual components, was moved into T234198.
Using an asset from node_modules works nicely out-of-the box.

Bug: T233791
Change-Id: Ie8d707e842a6b2e39c75e9a2bf5c54681c72e5f7