bridge: use data-bridge.css
So far data-bridge.css and data-bridge.app.css had no difference, but on
introducing assets I realized that the paths in data-bridge.app.css were
prefixed (../) which makes the files impossible to discover in the dist
folder.
Would be great to fully understand this but with a functional file at
hand this tries to unlock the button task.

Change-Id: If3a9349850b70111aa55d1f9a0eb5a0f964d4e5d