TR: Fix difference in CI and local build environment
For a currently unknown reason build output varies
when wikibase core is visible to the build script
vs when it is not.

This patch ensure that wikibase is visible when building
using the recommended docker development environment.

It also removes node_modules/.cache before building which may
be polluted by any builds created where wikibase core
is not visible to the build script.

Change-Id: I7c88682cc8be03fe4115863e06dbe9bdfa5f5142