bridge: use external typescript eslint
use external (TS) eslint rules and adapt where violations are reported
after rules came into effect through the update to typescript-eslint 2.
https://github.com/typescript-eslint/typescript-eslint/releases/tag/v2.0.0

This intentionally drags the dependency to eslint-config-wikimedia a
little because 0.15.0 introduced reportUnusedDisableDirectives as a
config variable which is incompatible with eslint 5 on which multiple of
our (indirect) dependencies still rely - it's quite a tree and I'd like
to update this independently.

Change-Id: I8d09efabba40fec766902d284ee1485b27a74f4d