bridge: add wbdatabridgeconfig internal API
This module is necessary so that the client-side code (both in the
“Wikibase client” and “browser” sense) can discover the maximum length
of strings, and apply it client-side. Some more parts of the
configuration will likely be made available later.

The module is only available if the dataBridgeEnabled repo setting is
true, via the ApiQuery::moduleManager hook; in repo/Wikibase.php, where
other API modules are registered, we don’t have custom settings yet.

Bug: T235034
Change-Id: I85d390c8b8974de93e5ba659892b074c9aef6263