Remove use of internal mw.loader.require() function
This exists for debugging purposes only and may stop working in the future.
All ResourceLoader-provided execution contexts have a lexically bound
'require()' function already that is dedicated to that response and should
be used instead.

For non-RL contexts such as inline HTML, there either are no dependencies
that you may rely on (and thus have no access to 'require()'), or, if
`mw.loader.using()` is used inline (like in this commit), then that function
provides acccess to the 'require' function.

<https://doc.wikimedia.org/mediawiki-core/master/js/#!/api/mw.loader-method-require>
<https://doc.wikimedia.org/mediawiki-core/master/js/#!/api/mw.loader-method-using>

Change-Id: I09c07beacf370d0403bafcf9e49e89a232dc3bde