Make warning less noisy from ClientParserOutputDataUpdater
This was previously using wfLogWarning which emits a native PHP
error, collected in production in the general "error" channel,
which counts towards metrics that measure production stability,
fire Icinga alerts, and may abort deployments if elevated.

Due to containing variables, it was not deduplicated and hard
to filter out from Logstash dashboards.

Convert to PSR-3 warning message in the Wikibase channel,
and also use Monolog context fields to allow aggregate counting.

Bug: T183993
Change-Id: Ia71abf183f9337e493c107b116c26f25a0d16325