Mark partial block test broken temporarily in SpecialNewEntityTestCase
After Iec36876e9 throwing a UserBlockedError results in the
BlockErrorFormatter calling AbstractBlock::getReasonComment, which
returns the block reason as a CommentStoreComment, which is then
formatted appropriately.

getReasonComment must be overridden for the mock block in
testExceptionWhenUserBlockedOnNamespace, or the formatter will attempt
to format the default null return value and fail. This can only be
done after getReasonComment is introduced in Iec36876e9.

Change-Id: Ia5ef8069a565ea60f9fdcef42534546e0ea1e018