bridge: Create progressive disabled version of save button
The click handling is effectively disabled twice:
1: by making the element impossible to select via keyboard and letting
it ignore all mouse/touch interaction via pointer-events: none;
2: by not propagating the click event in the click handler

The second is needed to make the unit tests pass, because the test utils
doesn't actually simulate a click and thus doesn't take css into
account, but rather it triggers the given event, irrespective of whether
that would actually happen in a browser. Aside from that, this second
way of disabling also serves as a safeguard in case something was missed
in the first step.

Bug: T235055
Change-Id: I1f60e9dbfd06b425de291761d19e26a9affd69de