bridge: save revision data
This patch adds the target statement revision to store. Since it is a
specific demand of DataBridge it is added to the application layer of
the store. Therefore it adds one new mutation and alters the
BRIDGE_INIT action.
It access the data directly via the Application state to avoid
uneccessary overdo. In this regard it alters mockedStore, too.

Bug: T235052
Change-Id: I9b1bafa15d9cb1352670435882d89bed64e8128f