Replace old @see PHPDoc tags with current @inheritDoc syntax
We introduced the @see syntax a long, long time ago in this codebase. Back
then I had no idea something like @inheritDoc existed (it likely didn't).
The meaning is the same. The issue with the @see syntax is that it can
contain mistakes or get outdated, e.g. when a referenced method is moved
from one parent class to another. Or the @see reference can just be plain
wrong. @inheritDoc avoids the repetition.

Note I did not checked if the @see tags I'm touching in this patch pointed
to the correct parent class before! This might result in rare false
positives where a @see points to a method with the *same* name in an
*unrelated* class. I believe the risk of having such a false positive is
worth it.

Change-Id: Ibef89beee46e6eb8ef1c84b35c62008202a8f088