termbox: entity page restyling via skinStyles
Restyling of entity pages (as a delta on top of minerva), which was
implemented during the termbox project, so far was realized by using the
skin name to build strong(er) selectors. It is noteworthy that these
do not apply to the "term box" itself but to the page surrounding it.
TIL that skinStyles are a thing, which can solve the same challenge by
only loading certain styles under a particular skin - which nicely
complements the 'delta' approach (ignoring its other shortcomings).
The selectors now should be a bit shorter but otherwise the functional
difference is probably academic, given that these styles are applied on
target 'mobile' only (which is minerva).

Change-Id: I1044d9ea39a1890a75dab4496b528605b8a4c011